Move the make_policy function from tain_imitation.py to the policy ingredient.
