+++ src/imitation/scripts/train_imitation.py
@@ -4,11 +4,11 @@ import logging
 import os.path as osp
 import pathlib
 import warnings
-from typing import Any, Mapping, Optional, Sequence, Type, cast
+from typing import Any, Mapping, Optional, Sequence, cast
 
 import numpy as np
 from sacred.observers import FileStorageObserver
-from stable_baselines3.common import policies, utils, vec_env
+from stable_baselines3.common import vec_env
 
 from imitation.algorithms import bc as bc_algorithm
 from imitation.algorithms.dagger import SimpleDAggerTrainer
@@ -16,26 +16,17 @@ from imitation.data import rollout, types
 from imitation.scripts.config.train_imitation import train_imitation_ex
 from imitation.scripts.ingredients import demonstrations, environment, expert
 from imitation.scripts.ingredients import logging as logging_ingredient
-from imitation.scripts.ingredients import policy_evaluation
+from imitation.scripts.ingredients import policy, policy_evaluation
 
 logger = logging.getLogger(__name__)
 
 
-@train_imitation_ex.capture(prefix="policy")
-def make_policy(
-    venv: vec_env.VecEnv,
-    policy_cls: Type[policies.BasePolicy],
-    policy_kwargs: Mapping[str, Any],
-    agent_path: Optional[str],
-) -> policies.BasePolicy:
-    """Makes policy.
+@train_imitation_ex.capture
+def make_or_load_policy(venv: vec_env.VecEnv, agent_path: Optional[str]):
+    """Makes a policy or loads a policy from a path if provided.
 
     Args:
         venv: Vectorized environment we will be imitating demos from.
-        policy_cls: Type of a Stable Baselines3 policy architecture.
-            Specify only if policy_path is not specified.
-        policy_kwargs: Keyword arguments for policy constructor.
-            Specify only if policy_path is not specified.
         agent_path: Path to serialized policy. If provided, then load the
             policy from this path. Otherwise, make a new policy.
             Specify only if policy_cls and policy_kwargs are not specified.
@@ -43,27 +34,15 @@ def make_policy(
     Returns:
         A Stable Baselines3 policy.
     """
-    policy_kwargs = dict(policy_kwargs)
-    if issubclass(policy_cls, policies.ActorCriticPolicy):
-        policy_kwargs.update(
-            {
-                "observation_space": venv.observation_space,
-                "action_space": venv.action_space,
-                # parameter mandatory for ActorCriticPolicy, but not used by BC
-                "lr_schedule": utils.get_schedule_fn(1),
-            },
-        )
-    policy: policies.BasePolicy
-    if agent_path is not None:
+    if agent_path is None:
+        policy.make_policy(venv)
+    else:
         warnings.warn(
-            "When agent_path is specified, policy_cls and policy_kwargs are ignored.",
+            "When agent_path is specified, policy.policy_cls and policy.policy_kwargs "
+            "are ignored.",
             RuntimeWarning,
         )
-        policy = bc_algorithm.reconstruct_policy(agent_path)
-    else:
-        policy = policy_cls(**policy_kwargs)
-    logger.info(f"Policy network summary:\n {policy}")
-    return policy
+        return bc_algorithm.reconstruct_policy(agent_path)
 
 
 @train_imitation_ex.capture
@@ -72,7 +51,6 @@ def train_imitation(
     bc_train_kwargs: Mapping[str, Any],
     dagger: Mapping[str, Any],
     use_dagger: bool,
-    agent_path: Optional[str],
     _run,
     _rnd: np.random.Generator,
 ) -> Mapping[str, Mapping[str, float]]:
@@ -83,9 +61,7 @@ def train_imitation(
         bc_train_kwargs: Keyword arguments passed through to `BC.train()` method.
         dagger: Arguments for DAgger training.
         use_dagger: If True, train using DAgger; otherwise, use BC.
-        agent_path: Path to serialized policy. If provided, then load the
-            policy from this path. Otherwise, make a new policy.
-            Specify only if policy_cls and policy_kwargs are not specified.
+        _run: Sacred run object.
         _rnd: Random number generator provided by Sacred.
 
     Returns:
@@ -94,7 +70,7 @@ def train_imitation(
     custom_logger, log_dir = logging_ingredient.setup_logging()
 
     with environment.make_venv() as venv:
-        imit_policy = make_policy(venv, agent_path=agent_path)
+        imit_policy = make_or_load_policy(venv)
 
         expert_trajs: Optional[Sequence[types.Trajectory]] = None
         if not use_dagger or dagger["use_offline_rollouts"]:
